lg_app/cricket-agent-app/requirements.txt
python-dotenv
langchain
langchain-community
langgraph
# langchain-anthropic
# langchain-openai
# langchain-google-genai
# langchain-ollama
langchain-deepseek

#grandlaf is to draw the graph workflow
grandalf

psycopg[binary,pool]
langgraph-checkpoint-postgres

# langgraph-checkpoint-redis